--- a/hw10_program_optimization/new
+++ b/hw10_program_optimization/new
@@ -1,0 +1,11 @@
+goos: darwin
+goarch: amd64
+pkg: github.com/maraero/otus-go/hw10_program_optimization
+cpu: Intel(R) Core(TM) i7-8850H CPU @ 2.60GHz
+BenchmarkGetDomainStat-12    	 2115252	      5720 ns/op	    6330 B/op	      84 allocs/op
+BenchmarkGetDomainStat-12    	 2110527	      5723 ns/op	    6330 B/op	      84 allocs/op
+BenchmarkGetDomainStat-12    	 2101341	      5754 ns/op	    6330 B/op	      84 allocs/op
+BenchmarkGetDomainStat-12    	 2085957	      5736 ns/op	    6330 B/op	      84 allocs/op
+BenchmarkGetDomainStat-12    	 2076276	      5760 ns/op	    6330 B/op	      84 allocs/op
+PASS
+ok  	github.com/maraero/otus-go/hw10_program_optimization	89.453s