--- a/hw10_program_optimization/old
+++ b/hw10_program_optimization/old
@@ -1,0 +1,11 @@
+goos: darwin
+goarch: amd64
+pkg: github.com/maraero/otus-go/hw10_program_optimization
+cpu: Intel(R) Core(TM) i7-8850H CPU @ 2.60GHz
+BenchmarkGetDomainStat-12    	      79	 141307804 ns/op	136045595 B/op	 1700090 allocs/op
+BenchmarkGetDomainStat-12    	      76	 138358678 ns/op	136045593 B/op	 1700090 allocs/op
+BenchmarkGetDomainStat-12    	      75	 138232932 ns/op	136045588 B/op	 1700090 allocs/op
+BenchmarkGetDomainStat-12    	      79	 138161389 ns/op	136045590 B/op	 1700090 allocs/op
+BenchmarkGetDomainStat-12    	      78	 138460784 ns/op	136045590 B/op	 1700090 allocs/op
+PASS
+ok  	github.com/maraero/otus-go/hw10_program_optimization	57.621s