## Домашнее задание №10 «Оптимизация программы»

### Чек-лист студента ([Что это?](https://github.com/OtusGolang/home_work/wiki/%D0%9A%D0%BE%D0%BC%D0%BC%D0%B5%D0%BD%D1%82%D0%B0%D1%80%D0%B8%D0%B8-%D0%BA-%D1%87%D0%B5%D0%BA-%D0%BB%D0%B8%D1%81%D1%82%D1%83-%D1%81%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0))
- [x] Я перечитал задание после решения. [Зачем перечитывать задание?](https://github.com/OtusGolang/home_work/wiki/%D0%9A%D0%BE%D0%BC%D0%BC%D0%B5%D0%BD%D1%82%D0%B0%D1%80%D0%B8%D0%B8-%D0%BA-%D1%87%D0%B5%D0%BA-%D0%BB%D0%B8%D1%81%D1%82%D1%83-%D1%81%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0#user-content-%D0%97%D0%B0%D1%87%D0%B5%D0%BC-%D0%BF%D0%B5%D1%80%D0%B5%D1%87%D0%B8%D1%82%D1%8B%D0%B2%D0%B0%D1%82%D1%8C-%D0%B7%D0%B0%D0%B4%D0%B0%D0%BD%D0%B8%D0%B5)
- [x] Я запустил `go mod tidy`.
- [x] Я удалил `.sync` файл. [Зачем его удалять?](https://github.com/OtusGolang/home_work/wiki/%5B%D0%A1%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0%D0%BC%5D-%D0%9F%D1%80%D0%BE%D1%86%D0%B5%D1%81%D1%81-%D1%81%D0%B4%D0%B0%D1%87%D0%B8-%D0%94%D0%97#user-content-%D0%92%D0%B0%D1%80%D0%B8%D0%B0%D0%BD%D1%82-2)
- [x] Я написал бенчмарк и приложил результаты benchstat.

### Критерии оценки
- [ ] Пайплайн зелёный и нет попытки «обмануть» систему - 4 балла
- [ ] Добавлены юнит-тесты - до 3 баллов
- [ ] Понятность и чистота кода - до 3 баллов

#### Зачёт от 7 баллов

------------------------------------------------------------------------------------------------------

benchstat -delta-test none old new
name              old time/op    new time/op    delta
GetDomainStat-12     138ms ± 0%       0ms ± 0%  -100.00%

name              old alloc/op   new alloc/op   delta
GetDomainStat-12     136MB ± 0%       0MB ± 0%  -100.00%

name              old allocs/op  new allocs/op  delta
GetDomainStat-12     1.70M ± 0%     0.00M ± 0%  -100.00%

Before
goos: darwin
goarch: amd64
pkg: github.com/maraero/otus-go/hw10_program_optimization
cpu: Intel(R) Core(TM) i7-8850H CPU @ 2.60GHz
BenchmarkGetDomainStat-12    	      79	 141307804 ns/op	136045595 B/op	 1700090 allocs/op
BenchmarkGetDomainStat-12    	      76	 138358678 ns/op	136045593 B/op	 1700090 allocs/op
BenchmarkGetDomainStat-12    	      75	 138232932 ns/op	136045588 B/op	 1700090 allocs/op
BenchmarkGetDomainStat-12    	      79	 138161389 ns/op	136045590 B/op	 1700090 allocs/op
BenchmarkGetDomainStat-12    	      78	 138460784 ns/op	136045590 B/op	 1700090 allocs/op
PASS
ok  	github.com/maraero/otus-go/hw10_program_optimization	57.621s

After
goos: darwin
goarch: amd64
pkg: github.com/maraero/otus-go/hw10_program_optimization
cpu: Intel(R) Core(TM) i7-8850H CPU @ 2.60GHz
BenchmarkGetDomainStat-12        2124907              5634 ns/op            6330 B/op         84 allocs/op
BenchmarkGetDomainStat-12        2135946              5624 ns/op            6330 B/op         84 allocs/op
BenchmarkGetDomainStat-12        2116936              5631 ns/op            6330 B/op         84 allocs/op
BenchmarkGetDomainStat-12        2136614              5645 ns/op            6330 B/op         84 allocs/op
BenchmarkGetDomainStat-12        2130895              5640 ns/op            6330 B/op         84 allocs/op
PASS
ok      github.com/maraero/otus-go/hw10_program_optimization    88.742s